// frontend/src/modules/autoria/shared/types/autoria.ts
export interface AdContact {
  id?: number;
  type: string;
  value: string;
  is_visible: boolean;
  is_primary: boolean;
  note?: string;
}